requirements.txt
matplotlib
PyQt5
pySerial